pyproject.toml
[tool.poetry]
name = "sd-bot"
version = "0.1.0"
description = ""
authors = ["Saturdays.AI"]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
